feat: enhanced population, sky burgs, air routes, geojson exports
- Replace 2-stage burg placement with 6-tier hierarchy (capital, largePort,
  regionalCenter, marketTown, largeVillage, smallVillage, hamlet) with
  cultural spacing modifiers and tier-based population ranges
- Add flying/skyPort/altitude properties to Burg interface with dedicated
  skyburg icon group and Sky Realm state support
- Add air route generation between sky ports using Urquhart graph with
  purple dashed styling across all 12 style presets
- Fix state statistics double-counting and limit province centers to
  major settlements (capped at 20 per state)
- Create standalone GeoJSON export bookmarklet scripts for burgs, states,
  provinces, cultures, religions, routes, rivers, cells, markers, and
  bundled ZIP export
- Support up to 100K burgs (Uint16Array -> Uint32Array for cells.burg)
  and 500K cells (extended density map and slider ranges)
- Add nix flake for development dependencies

+++ public/modules/resample.js
@@ -79,7 +79,7 @@ window.Resample = (function () {
     pack.cells.pop = new Float32Array(pack.cells.i.length);
     pack.cells.culture = new Uint16Array(pack.cells.i.length);
     pack.cells.state = new Uint16Array(pack.cells.i.length);
-    pack.cells.burg = new Uint16Array(pack.cells.i.length);
+    pack.cells.burg = new Uint32Array(pack.cells.i.length);
     pack.cells.religion = new Uint16Array(pack.cells.i.length);
     pack.cells.province = new Uint16Array(pack.cells.i.length);
